+++ NEWS.md
@@ -1,6 +1,7 @@
 # densityratio 0.2.1
 
 * patch test files to work with the new upcoming ggplot release
+* update link readme
 
 # densityratio 0.2.0