+++ NEWS.md
@@ -1,3 +1,5 @@
+# densityratio (development version)
+
 # densityratio 0.2.1
 
 * patch test files to work with the new upcoming ggplot release