+++ NEWS.md
@@ -1,4 +1,4 @@
-# densityratio (development version)
+# densityratio 0.2.2
 
 * update `plot_bivariate()` to depend on `ggh4x` to remove empty panels, instead
 of a grob (such that it remains a ggplot) object